fix: Fix integration tests failure from setting a custom workspace
